nl2sql_azureai_universal/requirements-production.txt
# ========================================
# Production Requirements for NL2SQL Teams Bot
# ========================================

# Microsoft 365 Agents SDK
microsoft-agents-activity==0.4.0
microsoft-agents-hosting-core==0.4.0
microsoft-agents-hosting-aiohttp==0.4.0
microsoft-agents-authentication-msal==0.4.0

# Azure AI & OpenAI
azure-ai-projects==1.0.0b6
azure-ai-inference==1.0.0b6
azure-identity==1.19.0
openai==1.59.8

# Web Server
aiohttp>=3.11.11
gunicorn==23.0.0

# Database
pyodbc==5.2.0

# Utilities
python-dotenv==1.1.1
pydantic>=2.10.4

# Dependencies pulled in by above packages
pyjwt>=2.10.1
isodate>=0.6.1
azure-core>=1.30.0
msal>=1.31.1
requests>=2.32.3
cryptography>=44.0.0
